% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{barcharts}
\alias{barcharts}
\title{Bar Charts.}
\usage{
barcharts(message, width = NULL, height = NULL)
}
\arguments{
\item{message}{A dataframe.}

\item{width, height}{nada.}
}
\value{
A bar chart based on ECharts 3.0.
}
\description{
Create a bar chart based on ECharts 3.0.
}

